requirements.txt
accelerate==1.3.0
bitsandbytes==0.45.1
diffusers==0.22.0
einops==0.8.0
face_alignment==1.4.1
facenet_pytorch==2.6.0
huggingface_hub==0.28.1
insightface==0.7.3
ipdb==0.13.13
kornia==0.8.0
loguru==0.7.3
numpy==1.23.0
omegaconf==2.3.0
opencv_python==4.11.0.86
opencv_python_headless==4.11.0.86
packaging==24.2
pandas==2.2.3
Pillow==11.1.0
preprocess==2.0.0
PyYAML==6.0.2
PyYAML==6.0.2
safetensors==0.5.2
scipy==1.15.1
setuptools==75.8.0
skimage==0.0
torchfile==0.1.0
tqdm==4.67.1
transformers==4.42.4
wandb==0.19.6
xformers==0.0.29.post2
yacs==0.1.8
